homework 5. Картинки отображаются и изменяются.

+++ src/main/java/ru/skypro/homework/repository/ImageRepository.java
@@ -7,6 +7,6 @@ import ru.skypro.homework.entity.Image;
 @Repository
 public interface ImageRepository extends JpaRepository<Image,Integer> {
 
-    Image getReferenceById(int id);
+    Image getReferenceById(long id);
 
 }
